refactored HomePage and split trending for code clarity
